+++ src/collective/eeafaceted/collectionwidget/config.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+NO_COLLECTIONWIDGET_EXCEPTION_MSG = 'The element you are trying to get collection-link criterion has no collection widget!'
+NO_FACETED_EXCEPTION_MSG = 'The element you are trying to get collection-link criterion on is not faceted enabled!'